amqp10_SUITE: Don't ack in `no_ack` mode
Otherwise we get a "precondition failed" error and the channel is
closed. This error occurred rarely because the client also closes the
channel in parallel. So depending on the order both close messages are
handled, the client might be happy (which is the case most of the time).

(cherry picked from commit 7dc9ee4a15b54eb8195e9befa1f73ffabb8e36d9)